first implementation of redux, displaying the list of words one to five on the le page
