--- front/components/le/index.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router';
-import LangButton from '../common/langbutton';
-
-class LearnEnv extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>Hello LearnEnv</h1>
-        <Link to="/" style={styles}><LangButton text="hello" backgroundColor="red" /></Link>
-      </div>
-    );
-  }
-}
-
-const styles = {
-  textDecoration: "none",
-  color: "black"
-};
-
-export default LearnEnv;
