+++ front/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import WordsReducer from './reducer_words';
+
+const rootReducer = combineReducers({
+  words: WordsReducer
+});
+
+export default rootReducer;
