+++ front/reducers/reducer_words.js
@@ -1,0 +1,9 @@
+export default function() {
+  return [
+    {word: 'one'},
+    {word: 'two'},
+    {word: 'three'},
+    {word: 'four'},
+    {word: 'five'}
+  ]
+}
