+++ front/routes.js
@@ -3,7 +3,7 @@ import { Route, IndexRoute } from 'react-router';
 import App from './app';
 import Home from './components/home/index';
 import Admin from './components/admin/index';
-import LearnEnv from './components/le/index';
+import LearnEnv from './containers/le/index';
 import Settings from './components/settings/index';
 
 export default (
